+++ source/src/PyCallable.cpp
@@ -39,7 +39,6 @@ namespace
 Ostap::Functions::PyCallable::PyCallable 
 ( PyObject* callable , const bool /* ok */ ) 
   : m_callable  ( callable ) 
-  , m_arguments ( PyTuple_New ( 1 ) )  
 {
   //
   Ostap::Assert ( m_callable && PyCallable_Check ( m_callable ) , 
@@ -53,7 +52,6 @@ Ostap::Functions::PyCallable::PyCallable
 Ostap::Functions::PyCallable::PyCallable 
 ( const Ostap::Functions::PyCallable&  right ) 
   : m_callable  ( right.m_callable  ) 
-  , m_arguments ( PyTuple_New ( 1 ) )  
 {
   Py_XINCREF ( m_callable ) ;
 }
@@ -63,10 +61,8 @@ Ostap::Functions::PyCallable::PyCallable
 Ostap::Functions::PyCallable::PyCallable 
 ( Ostap::Functions::PyCallable&&  right ) 
   : m_callable  ( right.m_callable  ) 
-  , m_arguments ( right.m_arguments ) 
 {
   right.m_callable  = nullptr ;
-  right.m_arguments = nullptr ;
 }
 // ============================================================================
 // Destructor 
@@ -74,7 +70,6 @@ Ostap::Functions::PyCallable::PyCallable
 Ostap::Functions::PyCallable::~PyCallable() 
 {
   if  ( m_callable  ) { Py_DECREF ( m_callable  ) ; m_callable  = nullptr ; }
-  if  ( m_arguments ) { Py_DECREF ( m_arguments ) ; m_arguments = nullptr ; }
 }
 // ============================================================================
 // the actual evaluation of function
@@ -83,23 +78,35 @@ double Ostap::Functions::PyCallable::evaluate ( const  double x ) const
 {
   // 
   if  ( nullptr == m_callable || !PyCallable_Check( m_callable  ) ) 
-  {
-    PyErr_Print() ;
-    Ostap::throwException ( s_ERROR1 , s_METHOD2 , 511 ) ;
-  }
+    {
+      PyErr_Print() ;
+      Ostap::throwException ( s_ERROR1 , s_METHOD2 , 511 ) ;
+    }
+  //
+  PyObject* arguments = PyTuple_New ( 1 ) ;
+  if ( !arguments )
+    {
+      PyErr_Print () ;
+      Ostap::throwException ( "Can't create PyTuple"    ,
+			      "PyCallable::evaluate"    ,
+			      Ostap::StatusCode ( 541 ) ) ;
+      return 0 ;
+    }
   //
   PyObject* px =  PyFloat_FromDouble ( x ) ;
-  if ( 0 != PyTuple_SetItem ( m_arguments , 0 , px ) ) 
-  {
-    PyErr_Print () ;
-    Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
-  }
+  if ( 0 != PyTuple_SetItem ( arguments , 0 , px ) ) 
+    {
+      PyErr_Print () ;
+      Py_DECREF ( arguments ) ; arguments = nullptr ;
+      Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
+    }
   //
-  PyObject* result = PyObject_CallObject ( m_callable , m_arguments ) ;
+  PyObject* result = PyObject_CallObject ( m_callable , arguments ) ;
+  //
+  Py_XDECREF ( arguments ) ;
   //
   return result_to_double ( result , s_METHOD2 ) ;
 }
-
 
 // ============================================================================
 // constructor from the callable object 
@@ -107,7 +114,6 @@ double Ostap::Functions::PyCallable::evaluate ( const  double x ) const
 Ostap::Functions::PyCallable2::PyCallable2 
 ( PyObject* callable , const bool /* ok */ ) 
   : m_callable  ( callable ) 
-  , m_arguments ( PyTuple_New ( 2 ) )  
 {
   //
   Ostap::Assert ( m_callable && PyCallable_Check ( m_callable ) , 
@@ -121,7 +127,6 @@ Ostap::Functions::PyCallable2::PyCallable2
 Ostap::Functions::PyCallable2::PyCallable2 
 ( const Ostap::Functions::PyCallable2&  right ) 
   : m_callable  ( right.m_callable  ) 
-  , m_arguments ( PyTuple_New ( 2 ) )  
 {
   Py_XINCREF ( m_callable ) ;
 }
@@ -131,10 +136,8 @@ Ostap::Functions::PyCallable2::PyCallable2
 Ostap::Functions::PyCallable2::PyCallable2 
 ( Ostap::Functions::PyCallable2&&  right ) 
   : m_callable  ( right.m_callable  ) 
-  , m_arguments ( right.m_arguments ) 
 {
   right.m_callable  = nullptr ;
-  right.m_arguments = nullptr ;
 }
 // ============================================================================
 // Destructor 
@@ -142,7 +145,6 @@ Ostap::Functions::PyCallable2::PyCallable2
 Ostap::Functions::PyCallable2::~PyCallable2() 
 {
   if  ( m_callable  ) { Py_DECREF ( m_callable  ) ; m_callable  = nullptr ; }
-  if  ( m_arguments ) { Py_DECREF ( m_arguments ) ; m_arguments = nullptr ; }
 }
 // ============================================================================
 // the actual evaluation of function
@@ -158,21 +160,35 @@ double Ostap::Functions::PyCallable2::evaluate
     Ostap::throwException ( s_ERROR1 , s_METHOD2 , 511 ) ;
   }
   //
+  PyObject* arguments = PyTuple_New ( 2 ) ;
+  if ( !arguments )
+    {
+      PyErr_Print () ;
+      Ostap::throwException ( "Can't create PyTuple"    ,
+			      "PyCallable2::evaluate"   ,
+			      Ostap::StatusCode ( 542 ) ) ;
+      return 0 ;
+    }
+  //
   PyObject* px =  PyFloat_FromDouble ( x ) ;
-  if ( 0 != PyTuple_SetItem ( m_arguments , 0 , px ) ) 
-  {
-    PyErr_Print () ;
-    Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
-  }
+  if ( 0 != PyTuple_SetItem ( arguments , 0 , px ) ) 
+    {
+      PyErr_Print () ;
+      Py_DECREF ( arguments ) ; arguments = nullptr ;
+      Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
+    }
   //
   PyObject* py =  PyFloat_FromDouble ( y ) ;
-  if ( 0 != PyTuple_SetItem ( m_arguments , 1 , py ) ) 
-  {
-    PyErr_Print () ;
-    Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
-  }
+  if ( 0 != PyTuple_SetItem ( arguments , 1 , py ) ) 
+    {
+      PyErr_Print () ;
+      Py_DECREF ( arguments ) ; arguments = nullptr ;
+      Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
+    }
   //
-  PyObject* result = PyObject_CallObject ( m_callable , m_arguments ) ;
+  PyObject* result = PyObject_CallObject ( m_callable , arguments ) ;
+  //
+  Py_XDECREF ( arguments ) ;
   //
   return result_to_double ( result , s_METHOD2 ) ;
 }
@@ -184,7 +200,6 @@ double Ostap::Functions::PyCallable2::evaluate
 Ostap::Functions::PyCallable3::PyCallable3 
 ( PyObject* callable , const bool /* ok */ ) 
   : m_callable  ( callable ) 
-  , m_arguments ( PyTuple_New ( 3 ) )  
 {
   //
   Ostap::Assert ( m_callable && PyCallable_Check ( m_callable ) , 
@@ -198,7 +213,6 @@ Ostap::Functions::PyCallable3::PyCallable3
 Ostap::Functions::PyCallable3::PyCallable3 
 ( const Ostap::Functions::PyCallable3&  right ) 
   : m_callable  ( right.m_callable  ) 
-  , m_arguments ( PyTuple_New ( 3 ) )  
 {
   Py_XINCREF ( m_callable ) ;
 }
@@ -208,10 +222,8 @@ Ostap::Functions::PyCallable3::PyCallable3
 Ostap::Functions::PyCallable3::PyCallable3 
 ( Ostap::Functions::PyCallable3&&  right ) 
   : m_callable  ( right.m_callable  ) 
-  , m_arguments ( right.m_arguments ) 
 {
   right.m_callable  = nullptr ;
-  right.m_arguments = nullptr ;
 }
 // ============================================================================
 // Destructor 
@@ -219,7 +231,6 @@ Ostap::Functions::PyCallable3::PyCallable3
 Ostap::Functions::PyCallable3::~PyCallable3() 
 {
   if  ( m_callable  ) { Py_DECREF ( m_callable  ) ; m_callable  = nullptr ; }
-  if  ( m_arguments ) { Py_DECREF ( m_arguments ) ; m_arguments = nullptr ; }
 }
 // ============================================================================
 // the actual evaluation of function
@@ -236,32 +247,46 @@ double Ostap::Functions::PyCallable3::evaluate
     Ostap::throwException ( s_ERROR1 , s_METHOD2 , 511 ) ;
   }
   //
+  PyObject* arguments = PyTuple_New ( 3 ) ;
+  if ( !arguments )
+    {
+      PyErr_Print () ;
+      Ostap::throwException ( "Can't create PyTuple"    ,
+			      "PyCallable3::evaluate"   ,
+			      Ostap::StatusCode ( 543 ) ) ;
+      return 0 ;
+    }
+  //
   PyObject* px =  PyFloat_FromDouble ( x ) ;
-  if ( 0 != PyTuple_SetItem ( m_arguments , 0 , px ) ) 
+  if ( 0 != PyTuple_SetItem ( arguments , 0 , px ) ) 
   {
     PyErr_Print () ;
+    Py_DECREF ( arguments ) ; arguments = nullptr ;
     Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
   }
   //
   PyObject* py =  PyFloat_FromDouble ( y ) ;
-  if ( 0 != PyTuple_SetItem ( m_arguments , 1 , py ) ) 
+  if ( 0 != PyTuple_SetItem ( arguments , 1 , py ) ) 
   {
     PyErr_Print () ;
+    Py_DECREF ( arguments ) ; arguments = nullptr ;
     Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
   }
   //
   PyObject* pz =  PyFloat_FromDouble ( z ) ;
-  if ( 0 != PyTuple_SetItem ( m_arguments , 2 , pz ) ) 
+  if ( 0 != PyTuple_SetItem ( arguments , 2 , pz ) ) 
   {
     PyErr_Print () ;
+    Py_DECREF ( arguments ) ; arguments = nullptr ;
     Ostap::throwException ( s_ERROR2  , s_METHOD2 , 512 ) ;
   }
   //
-  PyObject* result = PyObject_CallObject ( m_callable , m_arguments ) ;
+  PyObject* result = PyObject_CallObject ( m_callable , arguments ) ;
+  //
+  Py_XDECREF ( arguments ) ;
   //
   return result_to_double ( result , s_METHOD2 ) ;
 }
-
 
 // ============================================================================
 //                                                                      The END